+++ src/models/catalog/catalog.js
@@ -72,4 +72,17 @@ export class Catalog extends Model {
       trx,
     );
   }
+
+  /**
+   * Returns vocabulary data.
+   * @method getVocabulary
+   * @param {Object} req Request object.
+   * @returns {Object} JSON object.
+   */
+  getVocabulary(req) {
+    return {
+      title: req.i18n(this.Title),
+      token: this.Title,
+    };
+  }
 }